requirements.txt
gradio>=4.0.0
huggingface_hub>=0.19.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
# Core dependencies
pyyaml>=6.0
python-dotenv>=0.19.0
pathlib>=1.0.1
typing-extensions>=4.0.0

# Voice services
azure-cognitiveservices-speech>=1.20.0
google-cloud-texttospeech>=2.10.0
google-cloud-speech>=2.12.0
openai>=1.0.0

# ML and audio processing
torch>=2.0.0
transformers>=4.20.0
librosa>=0.9.0
numpy>=1.21.0
sounddevice>=0.4.6

# Web interface
gradio>=3.50.0
requests>=2.31.0
flask>=2.0.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0